+++ src/com/intellij/idea/plugin/hybris/system/bean/model/Bean.java
@@ -63,6 +63,7 @@ public interface Bean extends DomElement, AbstractPojo {
      */
     @NotNull
     @Attribute(EXTENDS)
+    @Convert(BSBeanClassResolvingConverter.class)
     GenericAttributeValue<String> getExtends();